+++ src/cmd/mod.rs
@@ -2,6 +2,7 @@ use clap::Parser;
 
 pub(super) mod create;
 pub(super) mod delete;
+pub(super) mod enable;
 pub(super) mod generate;
 pub(super) mod get;
 pub(super) mod history;
@@ -9,8 +10,6 @@ pub(super) mod initialize;
 pub(super) mod list;
 pub(super) mod login;
 pub(super) mod logout;
-pub(super) mod use_identity_provider;
-pub(super) mod use_storage_provider;
 pub(super) mod version;
 
 #[derive(Debug, Parser)]
@@ -22,6 +21,10 @@ pub enum Subcommand {
     /// Delete a single manageable model
     #[command(name = "delete")]
     Delete(delete::Opts),
+
+    /// Enable a feature flag
+    #[command(name = "enable")]
+    Enable(enable::Opts),
 
     /// Generate various things like XIDs, GUIDs, Models, Controllers, Resources, and Services
     #[command(name = "generate", alias = "g")]
@@ -50,12 +53,6 @@ pub enum Subcommand {
     /// Remove your access token (session) from the config file in your home dir
     #[command(name = "logout")]
     Logout(logout::Opts),
-
-    #[command(name = "use-identity-provider", alias = "useIdentityProvider")]
-    UseIdentityProvider(use_identity_provider::Opts),
-
-    #[command(name = "use-storage-provider", alias = "useStorageProvider")]
-    UseStorageAdapter(use_storage_provider::Opts),
 
     /// Display the build version
     #[command(name = "version")]